I fucked git up, as my last commit can prove. Fixed that

+++ JNF_NEAT/neural_network_trainer.h
@@ -24,7 +24,7 @@ class NeuralNetworkTrainer {
 
 		~NeuralNetworkTrainer() = default;
 
-		void TrainUntilFitnessEquals(size_t fitnessToReach);
+		void TrainUntilFitnessEquals(double fitnessToReach);
 		void TrainUntilGenerationEquals(unsigned int generationsToTrain);
 
 		TrainedNeuralNetwork GetTrainedNeuralNetwork();
